Migrate auth.js to TS
